wrong output

+++ Assignments/P06/10263/main.cpp
@@ -8,14 +8,14 @@
 
 using namespace std;
 
-pair<double, double> closestPoint(point a, point b, point p){
+pair<float, float> closestPoint(point a, point b, point p){
     point ap = make_pair(p.first - a.first, p.second - a.second);
     point ab = make_pair(b.first - a.first, b.second - a.second);
 
-    double abSquare = pow(ab.first, 2) + pow(ab.second, 2);
-    double dotProd = ap.first * ab.first + ap.second * ab.second;
+    float abSquare = pow(ab.first, 2) + pow(ab.second, 2);
+    float dotProd = ap.first * ab.first + ap.second * ab.second;
 
-    double t = dotProd / abSquare;
+    float t = dotProd / abSquare;
     return(make_pair(a.first + ab.first*t, a.second + ab.second*t));
 }
 
@@ -30,7 +30,7 @@ int main(){
 
 
     point M, temp;
-    pair<double, double> out;
+    pair<float, float> out;
     int N, closest, dis;
     vector<pair<int, int> > segments;
     while(input >> M.first){
